refractor upgrade job for re-usablility
