--- pkg/upgrade/upgrade_vars.go
@@ -1,36 +1,0 @@
-/*
-Copyright 2020-2021 The OpenEBS Authors
-
-Licensed under the Apache License, Version 2.0 (the "License");
-you may not use this file except in compliance with the License.
-You may obtain a copy of the License at
-
-    http://www.apache.org/licenses/LICENSE-2.0
-
-Unless required by applicable law or agreed to in writing, software
-distributed under the License is distributed on an "AS IS" BASIS,
-WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-See the License for the specific language governing permissions and
-limitations under the License.
-*/
-package upgrade
-
-// UpgradeOpts are the upgrade options that are provided
-// with the CLI flags
-type UpgradeOpts struct {
-	CasType     string
-	ToVersion   string
-	ImagePrefix string
-	ImageTag    string
-}
-
-type UpgradeJobCfg struct {
-	fromVersion        string
-	toVersion          string
-	namespace          string
-	pvNames            []string
-	backOffLimit       int32
-	serviceAccountName string
-	logLevel           int32
-	additionalArgs     []string
-}
